Move external tests to lib.rs
